Commit initial start on a failure mode test
